Abstracted calculateAverage and request to API into reusable functions outside of the changeProduct action creator
