+++ client/store/helpers/calculateAverage.js
@@ -1,0 +1,7 @@
+
+// This is a placeholder for the makeAVERAGE function
+function getAverage () {
+  return 3;
+}
+
+export default getAverage;
